fixes #6:
With the `myHDF5Plugin` class the info is in the event document, but in
two parts (file_name & file_number; full_file_name, a waveform string,
is not rendered ... yet)

The code (`interlace.tomo.FrameNotifier.event()`) shows how to access
the full name of the frame file from within the event handler.

BUT, this method completely ignores the filestore and its handling.
That will be an upgrade for the next phase.